Follow up changes for swact check to work
This commit is an addition to the proxy API [1], to added semantic check
for USM upgrade to restrict host-swact and host-unlock.

This commit also removed the check that prevents upgrade and downgrade
when new hardware is detected during upgrade procedure.

Test Plan:

PASS: run the upgrade on DX with USM, observed the restrictions take place

Task: 49798
Story: 2010676

[1] https://review.opendev.org/c/starlingx/config/+/914974

Change-Id: I8ded9faf7691ce849d51ef39f7598f287c6f1ca4
Signed-off-by: Bin Qian <bin.qian@windriver.com>