In SX mark floating IPs as deprecated in dual-stack
With the dual-stack feature the system now can have 2 floating IPs per
network. In non-SX systems the floating IPs are managed by SM, but not
in AIO-SX, this is done via puppet, and it requires to mark floating
addresses as deprecated.

This change can now export IPv4 and IPv6 addresses to the
"platform::network::addresses::address_config" variable

Test Plan
[PASS] install AIO-SX and check if floating IPs have the correct
       flags
[PASS] in the installation configure dual-stack and check if floating
       IPs have the correct flags

Story: 2011027
Task: 49887

Change-Id: I48d727a05ef46105f984848f24f21f087efc64e2
Signed-off-by: Andre Kantek <andrefernandozanella.kantek@windriver.com>