Create update_users, create_roles and create_projects methods
This commit creates the update_users, create_roles
and create_projects methods in
openstack_config_endpoints.py, which is required in [1]
in order to set the required services, roles and users
during keystone bootstrap.

[1]: https://review.opendev.org/c/starlingx/ansible-playbooks/+/915284

Test plan:
Note that all of the test cases were performed with the changes from
[1].
1. PASS: Verify the services, roles and users were created after
the keystone bootstrap.
2. PASS: Validate that the admin and sysinv users have the
ignore_lockout_failure_attempts set to true.
3. PASS: Validate the sql dump of the keystone database generated in
a subcloud deployment in relation to the one generated before the
changes from [1].

Story: 2011035
Task: 49966

Change-Id: I5be50bec1174a451d11e4dbc2eff0b01fc182576
Signed-off-by: Raphael Lima <Raphael.Lima@windriver.com>