Dual-stack: generate pxeboot label without address ID
mtcClient needs that the pxeboot interface config filename to contain
the label "[network-id]" to find the pxeboot address by reading the
file in /etc/network/interfaces.d/.

During dual-stack development the format [net-id]-[addr-id] was
adopted to differentiate files for each protocol (ipv4 and IPv6) in
the same network.

Since this broke mtcClient operation, we keeping the previous format
to the pxeboot network, since it does not support dual-stack.

Test Plan
[PASS] Install AIO-DX with pxeboot, management and cluster-host on
       the same ethernet port and verify:
       - the system is up with no alarms.
       - the pxeboot config file does not contain "-N" suffix
[PASS] Install AIO-DX with pxeboot, management and cluster-host on
       the same bonding port and verify:
       - the system is up with no alarms.
       - the pxeboot config file does not contain "-N" suffix
[PASS] Install AIO-DX with pxeboot on the ethernet port, and
       management and cluster-host on a vlan port on top of it and
       verify:
       - the system is up with no alarms.
       - the pxeboot config file does not contain "-N" suffix

Story: 2011027
Task: 49919

Change-Id: I42ab55d15c0df7d6a14377278b2f7624e83cb836
Signed-off-by: Andre Kantek <andrefernandozanella.kantek@windriver.com>