Fix system-local-ca ca.crt for upgrades
Included code in the upgrade healthy check to prevent the usage of
ICAs in the 'ca.crt' field of the 'system-local-ca' secret.

Also, included code to search through the trusted bundle for the RCA
if 'ca.crt' is not filled in the secret or if it is incorrect.

Test plan:
PASS: Perform 'system health-query-upgrade', verify that an error is
      shown if the 'ca.crt' data of system-local-ca secret is filled
      with a copy of the ICA.
PASS: Updated platform certificates in stx 9. Verified that
      system-local-ca secret does not possess the field 'ca.crt'.
      Perform upgrade from stx 9.0 (AIO-SX).
      Verified that the 'ca.crt' field was filled with the correct RCA
      cert.

Story: 2009811
Task: 50017

Change-Id: Ia3603b25d6d730a3465026b6b5291761d068613a
Signed-off-by: Marcelo Loebens <Marcelo.DeCastroLoebens@windriver.com>