Increase timeout of 95-watch-apps-upgrade.sh by 50%
A long term solution to avoid apps not updating in time
would require significant changes to the upgrade
system, so a simple change is being made in the short
term to see if some additional time will help.  This
script only fails intermittently, so testing is simple
in nature.

Test Plan:
PASS: Upgrade from stx8 to stx9 without errors.

Closes-Bug: 2064315

Change-Id: If2571837b005e604a1001412401e35b8ce711867
Signed-off-by: Reed, Joshua <Joshua.Reed@windriver.com>