Enable optional host-fs ceph filesystem
This adds an optional ceph-lv filesystem to the cgts-vg volume group
which will be used to establish ceph specific storage on controllers and
workers. The presence of this filesystem on a particular host will be
used to determine where containerized Ceph services will be provisioned
via Rook.

Additional followup commits will be submitted that adds state
information for the filesystem and institutes semantic checks limiting
when this filesystem can be added.

Other minor related changes:
 - Rename is_filesystem_enabled() to is_host_filesystem_enabled() to
   differentiate between controller and host filesystems
 - Change FS_CREATION_ALLOWED to HOSTFS_CREATION_ALLOWED to
   differentiate between controller and host filesystems

Test Plan:
 - PASS: Install AIO-SX with bare-metal Ceph
 - PASS: Install AIO-DX with bare-metal Ceph
 - PASS: Install 2+2 with bare-metal Ceph
 - PASS: add/remove optional ceph host-fs on controller-0/1
 - PASS: add/remove optional ceph host-fs on worker

Change-Id: Iba08543374281922e4e1f44ec9bb3326f7a22c08
Story: 2011117
Task: 50036
Depends-On: https://review.opendev.org/c/starlingx/stx-puppet/+/918449
Signed-off-by: Robert Church <robert.church@windriver.com>