Fix 'sysinv-helm create-fluxcd-app-overrides' command
The 'sysinv-helm create-fluxcd-app-overrides <app_name>
<namespace>' is a command that allows a user to generate helm
override values for a helm chart independently of an application
upload/apply. This command is useful for testing.

It was broken due to a cache cleaning done in apps_metadata_dict
that mistakenly added a different key to the one being used.

The command correction is done by changing the dictionary's key
set to the correct key.

Test Plan:
PASS: Run "sysinv-helm create-fluxcd-app-overrides
      /home/sysadmin oidc-auth-apps kube-system"
      commamd and generate files with success.
PASS: Run "sysinv-helm create-fluxcd-app-overrides
      /home/sysadmin/ cert-manager cert-manager" command and
      generate files with success.

Closes-Bug: 2060864

Change-Id: If5aa2bc96577811182ed0fd326c55b229410c4ff
Signed-off-by: David Bastos <david.barbosabastos@windriver.com>