Store IPSec cert and keys in LUKS filesystem
This commit updated ipsec-client to store certs and keys in LUKS
filesystem. The encrypted filesystem take some time to mount and
decrypted, so a script was created to check if the filesystem is
ready before start ipsec.

Test Plan:
PASS: Bootstrap, install and unlock a DX system with a worker. Wait
      until system reboots and verify unlocked enable available status.
      Then, observe IPSec is enabled, SAs are established and the
      symbolic links were created in the directories /etc/swanctl/x509
      and /etc/swanctl/private pointing to file in the directory
      /var/luks/stx/luks_fs/ipsec/{certs,keys}.

Story: 2010940
Task: 50086

Change-Id: Ifccd747bd1db8f565d4744d99d94a61a22d5890e
Signed-off-by: Leonardo Mendes <Leonardo.MendesSantana@windriver.com>