Eliminate file check in Armada required function.
Previously the function was basing a decision whether
or not Armada is required soley on whether or not
the folder /opt/platform/armada exists or not.  There
might be the condition where an Armada application was
installed on a lower version of STX, then upgraded to
FluxCD, and the old armada manifests left over.  This
change seeks to make the determination about Armada
being required solely upon the helm2 cli showing a release
and if a pod existing in the armada namespace existing with
the "application=armada" label.

Test Plan:
PASS: Upgrade activation step between stx9.0 and future
      stx10.0 and 76 script passes.
PASS: Force install an old armada application to check
      if the 76 script leaves armada in place. In this
      case the Armada pod is detected and the script
      exit and activation fails as it should.

Closes-Bug:2065320

Change-Id: I50dabe843549f7f84522c2a61056560c5c084da5
Signed-off-by: Joshua Reed <joshua.reed@windriver.com>