Auto-install system-local-ca RCA as trusted
Use cert-mon to monitor and install system-local-ca's RCA
as a trusted CA upon changes in the secret.

This RCA anchors the platform certificates (OpenLDAP, Docker
Registry, REST API/GUI and such), and needs to be trusted
by the platform.

A known issue with OpenSSL results in problems when a new
RCA with the same subject as existing one is added to the
trusted bundle. Cert-mon will remove an existing RCA if
it has the same subject as the new one.

Test plan:
PASS: Run tox to validate the unit tests.
PASS: Bootstrap AIO-SX and DC + SX subcloud.
PASS: Run system-local-ca update.
PASS: Manually replaced system-local-ca 'ca.crt' with a RCA
      w/ same subject. Observed that cert-mon uninstalls the
      old RCA and installs the new one.

Story: 2009811
Task: 50275

Change-Id: I5a9cd1ad4063e24d8f5b976e922cc980aba2f612
Signed-off-by: Marcelo Loebens <Marcelo.DeCastroLoebens@windriver.com>