Point patch current health check to use USM
Starting on stx-11, patching will be managed by the
USM feature, so all patch related health-checks and
information should be gathered via USM APIs.

This commit changes the patch current health checks
to point to USM APIs. However, there is not an endpoint
that returns per-host data, and deploy host-list only
returns data if there is a deployment in progress, so
there will be a followup commit to change this logic
again if an endpoint is created to return data as the
legacy sw-patch query-hosts command.

This commit does not affect legacy upgrades, since
on legacy upgrade the health-check code runs on the
FROM release code base.

Test Plan
PASS: run 'deploy precheck' and verify the output
PASS: run 'deploy precheck' on a system with [Fail]
      patch current health check, apply the fix and
      verify the output

Story: 2010676
Task: 50107

Signed-off-by: Heitor Matsui <heitorvieira.matsui@windriver.com>
Change-Id: I69c630468543ac99d8ba79f21646573a91140616