Improvement log error indicating namespace not found
Namespace not found error message is shown in sysinv.log when an application that has a specific namespace is uploaded.

This log is expected and has no impact on the system. This
happens due to the command "kubectl -k <manifest_dir>
--dry-run=server". If the application has a specific namespace that
the system does not yet have, the command shows this error. But
there is nothing to worry about, as the namespace is created within
the apply process.

To avoid misinterpretations, this log has been suppressed.

Test Plan:
PASS: Add the new log to kube_app.py and restart sysinv.
PASS: During the application upload process, the log that says the
      namespace was not found no longer appears.

Closes-Bug: 2067070

Change-Id: I6e56c338c9789c686381c330be65ec109a3f0b50
Signed-off-by: David Bastos <david.barbosabastos@windriver.com>