Improve robustness of ipsec-config's ocf script
This update made various improvments to the ipsec-config's ocf script
to make it more resillient to error conditions. Particularly when the
swanctl.conf points to the wrong version of config file, the
ipsec-config service will correct it based on whether the host is
active or standby controller.

Test Plan:
PASS: In a DX system, swact controllers back and forth, verify after
      swact, swanctl.conf points to the correct version of swanctl
      config file, IPsec SAs established, both controllers are in
      unlocked|enabled|available state.
PASS: In a DX system, lock standby controller, then force reboot active
      controller by "reboot -f", verify that the swanctl.conf points to
      the swanct_standby.conf when it boots up, and then it is
      corrected to point to swanctl_active.conf by ipsec-config when SM
      services start up, also IPsec SAs are established between
      controllers. Eventually both controllers are in
      unlocked|enabled|available, system is stable with no reboot.
PASS: In a DX system, reboot both controllers, verify that the
      swanctl.conf points to the swanct_standby.conf when they boot up,
      then the symlink on active controller is corrected by
      ipsec-config when SM services start up, also IPsec SAs are
      established between controllers. Eventually both controllers are
      in unlocked|enabled|available, system is stable with no reboot.
PASS: In a DX system, manually link a wrong config file to
      swanctl.conf, verify that the ipsec-config service correct it.

Story: 2010940
Task: 50198

Change-Id: I9fc028eb7a0ed296f9dd47b92c2b53302e845ac3
Signed-off-by: Andy Ning <andy.ning@windriver.com>