Disable system upgrade command
This is to disable the system upgrade related commands.

In this commit, the 'system load' and 'system upgrade' cli sub
commands are removed. The REST API endpoints are set to
deprecated.

Test Plan:

PASS: build and deploy the iso
PASS: cli commands are removed
PASS: QA sanity check against the iso

Story: 2010676
Task: 50296
Change-Id: I58151da5c5cf319985623cacc4580a1a6be4c94d
Signed-off-by: junfeng-li <junfeng.li@windriver.com>